refactor: extracting bubbleAbort
