+++ src/AbortError.ts
@@ -1,0 +1,6 @@
+export class AbortError extends Error {
+  code = 'aborted'
+  constructor () {
+    super('aborted')
+  }
+}
